Introduced list adding

+++ Classes/Controllers/TodolistTaskController.php
@@ -2,30 +2,30 @@
 class TodolistTaskController extends Controller{
 
     public function addTask(){
-        $idTodolist = $_POST['idTodolist'];
-        $description = $_POST['description'];
-        $todolistTaskObj = new TodolistTask();
-        $todolistTaskObj->setIdTodolist($idTodolist);
-        $todolistTaskObj->setDescription($description);
-        $this->respond($todolistTaskObj->insert());
+        if(isset($_POST['idTodolist']) && isset($_POST['description'])){
+            $todolistTaskObj = new TodolistTask();
+            $todolistTaskObj->setIdTodolist($_POST['idTodolist']);
+            $todolistTaskObj->setDescription($_POST['description']);
+            $this->respond($todolistTaskObj->insert());
+        }
     }
 
     public function updateTask(){
-        $id = $_POST['id'];
-        $description = $_POST['description'];
-        $finished = $_POST['finished'];
-        $todolistTaskObj = new TodolistTask();
-        $todolistTaskObj->setId($id);
-        $todolistTaskObj->setDescription($description);
-        $todolistTaskObj->setFinished($finished);
-        $this->respond($todolistTaskObj->update());
+        if(isset($_POST['id']) && isset($_POST['description']) && isset($_POST['finished'])){
+            $todolistTaskObj = new TodolistTask();
+            $todolistTaskObj->setId($_POST['id']);
+            $todolistTaskObj->setDescription($_POST['description']);
+            $todolistTaskObj->setFinished($_POST['finished']);
+            $this->respond($todolistTaskObj->update());
+        }
     }
 
     public function deleteTask(){
-        $id = $_POST['id'];
-        $todolistTaskObj = new TodolistTask();
-        $todolistTaskObj->setId($id);
-        $this->respond($todolistTaskObj->delete());
+        if(isset($_POST['id'])){
+            $todolistTaskObj = new TodolistTask();
+            $todolistTaskObj->setId($_POST['id']);
+            $this->respond($todolistTaskObj->delete());
+        }
     }
 
 }
